Moved elementTypeToSelectorName to the helper module
