+++ lib/elementWrapper/elementWrapper.js
@@ -1,19 +1,4 @@
-const { firstElement, desc, prepareParameters } = require('./helper');
-
-function elementTypeToSelectorName(elementType) {
-  switch (elementType) {
-    case 'CustomSelector':
-      return '$';
-    case 'Checkbox':
-      return 'checkBox';
-    case 'Table':
-      return 'tableCell';
-    case 'Element':
-      return 'text';
-    default:
-      return elementType.charAt(0).toLowerCase() + elementType.slice(1);
-  }
-}
+const { firstElement, desc, prepareParameters, elementTypeToSelectorName } = require('./helper');
 
 const { descEvent } = require('../eventBus');
 let { getIfExists } = require('../elementSearch');
