+++ lib/elementWrapper/helper.js
@@ -85,9 +85,25 @@ const desc = (selector, query, tag, options) => {
   return description;
 };
 
+function elementTypeToSelectorName(elementType) {
+  switch (elementType) {
+    case 'CustomSelector':
+      return '$';
+    case 'Checkbox':
+      return 'checkBox';
+    case 'Table':
+      return 'tableCell';
+    case 'Element':
+      return 'text';
+    default:
+      return elementType.charAt(0).toLowerCase() + elementType.slice(1);
+  }
+}
+
 module.exports = {
   prepareParameters,
   getElementGetter,
   desc,
   firstElement,
+  elementTypeToSelectorName,
 };
